requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
greenlet==1.1.2
idna==3.3
PyYAML==6.0
requests==2.27.1
soupsieve==2.3.1
SQLAlchemy==1.4.31
urllib3==1.26.8
